refactor(qqwife): 将CD时间类型从float64改为int64
修改CD时间相关字段和函数参数类型，从float64改为int64
统一使用分钟作为时间单位，简化时间计算逻辑

+++ plugin/qqwife/favorSystem.go
@@ -86,7 +86,7 @@ func init() {
 				ctx.SendChain(message.Reply(ctx.Event.MessageID), message.Text("校验CD时时发生意外错误：", err))
 				return
 			}
-			ok, err := 民政局.判断礼物CD(gid, uid, groupInfo.GiftCDtime)
+			ok, err := 民政局.判断礼物CD(gid, uid, int64(groupInfo.GiftCDtime))
 			if err != nil {
 				ctx.SendChain(message.Reply(ctx.Event.MessageID), message.Text("校验CD时时发生意外错误：", err))
 				return
